src/ru/vasiliydz/javahomework/Account.java
package ru.vasiliydz.javahomework;

public class Account {
    private final long id;
    private double balance;

    public Account(long id) {
        this.id = id;
        this.balance = 0;
    }

    /**
     * Withdraws money from account balance
     *
     * @param amount amount of money to withdraw
     * @return true
     * if amount &gt 0 and (balance - amount) &gt 0,
     * otherwise returns false
     */
    public boolean withdraw(double amount) {
        if (amount <= 0) {
            return false;
        }
        if (balance - amount > 0) {
            balance -= amount;
            return true;
        }
        return false;
    }

    /**
     * Adds money to account balance
     *
     * @param amount amount of money to add on account
     * @return true if amount &gt 0, otherwise returns false
     */
    public boolean add(double amount) {
        if (amount > 0) {
            balance += amount;
            return true;
        }
        return false;
    }
}